Update project structure. Added Carthage support. Added Travis-CI support. Added Swift-linting
