+++ .swiftlint.yml
@@ -1,0 +1,61 @@
+disabled_rules: 
+  - file_length
+  - line_length
+  - cyclomatic_complexity
+  - type_body_length
+  - type_name
+  - nesting
+
+opt_in_rules:
+  - anyobject_protocol
+  - array_init
+  - closure_end_indentation
+  - closure_spacing
+  - contains_over_first_not_nil
+  - discouraged_optional_boolean
+  - empty_count
+  - empty_string
+  - empty_xctest_method
+  - explicit_init
+  - fallthrough
+  - first_where
+  - force_unwrapping
+  - inert_defer
+  - no_extension_access_modifier
+  - overridden_super_call
+  - prohibited_super_call
+  - redundant_nil_coalescing
+  - vertical_parameter_alignment_on_call
+  - pattern_matching_keywords
+  - fatal_error_message
+  - implicit_return
+  - implicitly_unwrapped_optional
+  - joined_default_parameter
+  - let_var_whitespace
+  - literal_expression_end_indentation
+  - lower_acl_than_parent
+  - modifier_order
+  - multiline_arguments
+  - multiline_function_chains
+  - multiline_parameters
+  - multiple_closures_with_trailing_closure
+  - notification_center_detachment
+  - number_separator
+  - object_literal
+  - operator_usage_whitespace
+  - override_in_extension
+  - private_action
+  - private_outlet
+  - redundant_type_annotation
+  - single_test_class
+  - sorted_imports
+  - sorted_first_last
+  - trailing_closure
+  - unavailable_function
+  - unneeded_parentheses_in_closure_argument
+  - yoda_condition
+
+function_body_length:
+  - 50 #warning
+
+reporter: "xcode"
