+++ .travis.yml
@@ -1,14 +1,12 @@
-# references:
-# * https://www.objc.io/issues/6-build-tools/travis-ci/
-# * https://github.com/supermarin/xcpretty#usage
+os: osx
+language: swift
+osx_image: xcode10.2
+xcode_project: JNGradientLabel.xcodeproj
 
-osx_image: xcode7.3
-language: objective-c
-# cache: cocoapods
-# podfile: Example/Podfile
-# before_install:
-# - gem install cocoapods # Since Travis is not always on latest version
-# - pod install --project-directory=Example
+env:
+  - DESTINATION='platform=iOS Simulator,OS=9.3,name=iPhone 5s'
+  - DESTINATION='platform=iOS Simulator,OS=10.0,name=iPhone 5s'
+  - DESTINATION='platform=iOS Simulator,name=iPhone XS Max'
+
 script:
-- set -o pipefail && xcodebuild test -enableCodeCoverage YES -workspace Example/GradientLabel.xcworkspace -scheme GradientLabel-Example -sdk iphonesimulator9.3 ONLY_ACTIVE_ARCH=NO | xcpretty
-- pod lib lint
+  - set -o pipefail && travis_retry xcodebuild -scheme JNGradientLabel -destination "$DESTINATION" -configuration Debug ONLY_ACTIVE_ARCH=YES
