--- Example/Podfile
@@ -1,5 +1,0 @@
-target 'JNGradientLabelExample' do
-  use_frameworks!
-  
-  pod 'JNGradientLabel', :path => '../'  
-end
