+++ JNGradientLabelExample/Images.xcassets/AppIcon.appiconset/Contents.json
